pyproject.toml
[project]
name = "lifeos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyyaml>=6.0.3",
    "questionary>=2.1.1",
    "rich>=14.3.2",
]
